backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
llama-index==0.9.15
llama-index-agent-openai==0.1.7
weaviate-client==3.25.3
friendli==0.4.0
boto3==1.34.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
